+++ src/Web/Event/EventTarget.js
@@ -6,6 +6,18 @@ export function eventListener(fn) {
   };
 }
 
+export function addEventListenerWithOptions(type) {
+  return function (listener) {
+    return function (options) {
+      return function (target) {
+        return function () {
+          return target.addEventListener(type, listener, options);
+        };
+      };
+    };
+  };
+}
+
 export function addEventListener(type) {
   return function (listener) {
     return function (useCapture) {